moved grpchelper methods into the grpc client implementation. Removed no longer used grpc helper methods
